251124/requirements.txt
# GNC-쿠팡 매칭 시스템 필수 패키지

# 데이터 처리
pandas>=2.0.0
openpyxl>=3.1.0

# 웹 크롤링
selenium>=4.0.0
requests>=2.28.0

# 이미지 처리
Pillow>=10.0.0
imagehash>=4.3.0

# AI 이미지 비교 (선택사항)
google-generativeai>=0.3.0

# 유틸리티
python-dateutil>=2.8.0